+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 ### To be released
+- InputFile Component (https://github.com/tony-dinh/react-toolkit/pull/10)
 
 ### v0.1.0
 - Initial Release